+++ crates/formatter/src/lib.rs
@@ -1,30 +1,25 @@
-pub mod cst;
-pub mod format;
-pub mod lexer;
-pub mod parser;
+mod format_doc;
+mod format_tree;
+mod lower_cst;
 
 use logos::Logos;
 use pretty::RcAllocator;
+use syntax::{lexer, parser};
 
-pub use cst::*;
-pub use format::*;
-pub use lexer::*;
-pub use parser::*;
+use crate::format_doc::FormatDoc;
+use crate::lower_cst::lower_toplevel;
 
-pub fn format_source(input: &str) -> String {
+pub fn format_source(input: &str, width: usize) -> String {
     let lexed = lexer::Token::lexer(input);
     let lexed_ok: Vec<_> = lexed.filter_map(|token| token.ok()).collect();
     let parsed = parser::parse(&mut lexed_ok.as_slice()).expect("failed to parse");
-    let lowered = parsed
-        .into_iter()
-        .map(cst::lower_toplevel)
-        .collect::<Vec<_>>();
+    let lowered = parsed.into_iter().map(lower_toplevel).collect::<Vec<_>>();
     let format_doc = lowered
         .iter()
-        .fold(format::FormatDoc::nil(&RcAllocator), |acc, fst| {
+        .fold(FormatDoc::nil(&RcAllocator), |acc, fst| {
             acc.append(fst.format(&RcAllocator))
         });
-    format_doc.pretty(80)
+    format_doc.pretty(width)
 }
 
 pub fn check_source(input: &str) -> bool {
