+++ crates/formatter/tests/fixtures/defun_singleline.pact
