+++ crates/syntax/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "syntax"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+logos = "0.14.2"
+pretty = "0.12.3"
+winnow = "0.6.20"
