+++ crates/syntax/src/lexer.rs
