+++ crates/syntax/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod cst;
+pub mod lexer;
+pub mod parser;
